Allow workflow to take relative path to output_dir and data_paths.
If relative path is used for data_paths, data is presumed to be relative to the
output_path, which is also the location of the parameters file. This might be a
bit dangerous, since a user can give the output_path relative to cwd (the
logical assumption), but the data_path (which is stored in parameters and is
not given at the command line) is relative to the location of the parameters.