Close logging explicitly in main
In Windows GUI, apparently closing the thread running the MacroecoDesktop
analysis on a parameter file does not automatically release the file lock,
which it appears to do on Mac and Linux. This fix allows the log file to
be removed during additional runs from within the Windows GUI without the
program being closed and reopened.